third_party/WebKit/LayoutTests/imported/web-platform-tests/html/semantics/interactive-elements/the-dialog-element/dialog-showModal-expected.txt

This is a testharness.js-based test.
FAIL dialog element: showModal() assert_false: expected false got undefined
PASS showModal() on a <dialog> that already has an open attribute throws an InvalidStateError exception 
PASS showModal() on a <dialog> not in a Document throws an InvalidStateError exception 
FAIL when opening multiple dialogs, only the newest one is non-inert assert_false: expected false got undefined
FAIL opening dialog without focusable children assert_equals: expected Element node <dialog id="d6" open=""></dialog> but got Element node <body><div id="log"></div>
<button id="b0">OK</button>
<d...
PASS opening dialog with multiple focusable children 
PASS opening dialog with multiple focusable children, one having the autofocus attribute 
Harness: the test ran to completion.
